работа над календарём, замена фото на svg
