+++ index.html
@@ -18,7 +18,7 @@
     <header class="header">
       <div class="header__nav">
         <div class="logo">
-          <img src="src/img/logo.png" alt="logo" />
+          <img src="src/img/logo.svg" alt="logo" />
         </div>
         <ul class="nav__items">
           <li class="nav__item">
@@ -27,9 +27,9 @@
           <li class="nav__item">
             <a href="#"><h2>Attractions</h2></a>
           </li>
-          <li class="nav__item"><img src="src/img/night.png" alt="day/night" /></li>
-          <li class="nav__item"><img src="src/img/account.png" alt="account" /></li>
-          <li class="nav__item"><img src="src/img/menu.png" alt="menu" /></li>
+          <li class="nav__item"><img src="src/img/night.svg" alt="day/night" /></li>
+          <li class="nav__item"><img src="src/img/account.svg" alt="account" /></li>
+          <li class="nav__item"><img src="src/img/menu.svg" alt="menu" /></li>
         </ul>
       </div>
       <div class="container">
@@ -81,8 +81,8 @@
                     <p>Adults</p>
                     <div class="count__people__block">
                       <svg id="less-people" width="30" height="30" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
-                        <rect x="0.5" y="0.5" width="29" height="29" fill="white" stroke="#CECECE" />
-                        <path d="M19.3242 16.2783H10.5176V14.9512H19.3242V16.2783Z" fill="#CECECE" />
+                        <rect x="0.5" y="0.5" width="29" height="29" fill="white" stroke="#3077C6" />
+                        <path d="M19.3242 16.2783H10.5176V14.9512H19.3242V16.2783Z" fill="#3077C6" />
                       </svg>
                       <p id="count__people">2</p>
                       <svg id="more-people" width="30" height="30" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -91,7 +91,6 @@
                           d="M16.2822 15.1357H19.8506V16.6738H16.2822V20.7168H14.6475V16.6738H11.0791V15.1357H14.6475V11.4004H16.2822V15.1357Z"
                           fill="#3077C6"
                         />
-                        <!-- #3077C6 -->
                       </svg>
                     </div>
                   </div>
@@ -109,7 +108,6 @@
                           d="M16.2822 15.1357H19.8506V16.6738H16.2822V20.7168H14.6475V16.6738H11.0791V15.1357H14.6475V11.4004H16.2822V15.1357Z"
                           fill="#3077C6"
                         />
-                        <!-- #3077C6 -->
                       </svg>
                     </div>
                   </div>
@@ -127,7 +125,6 @@
                           d="M16.2822 15.1357H19.8506V16.6738H16.2822V20.7168H14.6475V16.6738H11.0791V15.1357H14.6475V11.4004H16.2822V15.1357Z"
                           fill="#3077C6"
                         />
-                        <!-- #3077C6 -->
                       </svg>
                     </div>
                   </div>
@@ -193,25 +190,25 @@
           </div>
           <div class="advantages__items">
             <div class="advantages__item">
-              <img class="advantages__img" src="src/img/advant1.png" />
+              <img class="advantages__img" src="src/img/advant1.svg" />
               <div class="advantages__text">
                 <h2>Support 24/7</h2>
               </div>
             </div>
             <div class="advantages__item">
-              <img class="advantages__img" src="src/img/advant2.png" />
+              <img class="advantages__img" src="src/img/advant2.svg" />
               <div class="advantages__text">
                 <h2>Communicate directly</h2>
               </div>
             </div>
             <div class="advantages__item">
-              <img class="advantages__img" src="src/img/advant3.png" />
+              <img class="advantages__img" src="src/img/advant3.svg" />
               <div class="advantages__text">
                 <h2>Book online</h2>
               </div>
             </div>
             <div class="advantages__item">
-              <img class="advantages__img" src="src/img/advant4.png" />
+              <img class="advantages__img" src="src/img/advant4.svg" />
               <div class="advantages__text">
                 <h2>Real guest reviews</h2>
               </div>
@@ -329,7 +326,7 @@
               </div>
               <div class="signup__info">
                 <div class="signup__img">
-                  <img src="src/img/user.png" alt="user" />
+                  <img src="src/img/user.svg" alt="user" />
                 </div>
                 <div class="signup__content">
                   <div class="content__text">
@@ -597,6 +594,8 @@
         </div>
       </div>
     </footer>
+    <script type="module" src="src/js/people-info.js"></script>
+    <script type="module" src="src/js/calendar.js"></script>
     <script type="module" src="src/js/dataCalendar.js"></script>
     <script type="module" src="src/js/main.js"></script>
   </body>
